+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## xx.x.x
+* Fixed an issue where the 'reportFeedbackWidgetManually' function would await indefinitely on iOS
+* Resolved an issue where nonfatal exceptions were treated as fatal and vice versa
+
+* Underlying Android SDK version is 24.1.1
+* Underlying iOS SDK version is 24.1.0
+
+## xx.x.x-np
+* Fixed an issue where the 'reportFeedbackWidgetManually' function would await indefinitely on iOS
+* Resolved an issue where nonfatal exceptions were treated as fatal and vice versa
+
+* Underlying Android SDK version is 24.1.1
+* Underlying iOS SDK version is 24.1.0
+
 ## 24.1.1
 * Added a new metric for detecting whether or not a device has a hinge for Android